+++ netlogo-gui/src/main/workspace/ExtensionManager.scala
@@ -58,7 +58,7 @@ import org.nlogo.nvm.{ ExtensionManager => NvmExtensionManager }
 object ExtensionManager {
 
   case class ExtensionData(extensionName: String, fileURL: URL, prefix: String, classManagerName: String, version: Option[String], private val modifiedRaw: Long) {
-    val modified = 1000 * Math.round(modifiedRaw / 1000)
+    val modified = 1000 * Math.round(modifiedRaw.toFloat / 1000)
   }
 
   class JarContainer(val jarClassLoader: ClassLoader, data: ExtensionData) {
